+++ lib/https.dart
@@ -1,13 +1,18 @@
 // Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
+
+/// Node HTTPS module bindings.
+///
+/// Use top-level [https] object to access this module functionality.
+/// To create HTTPS agent use [createHttpsAgent].
 @JS()
-library node_interop.bindings.https;
+library node_interop.https;
 
 import 'dart:js_util';
 
 import 'package:js/js.dart';
 
-import 'globals.dart';
+import 'node.dart';
 import 'http.dart';
 import 'tls.dart';
 
@@ -19,25 +24,18 @@ export 'http.dart'
         IncomingMessage,
         ServerResponse;
 
+HTTPS get https => require('https');
+
 /// Convenience method for creating instances of "https" module's Agent class.
 ///
 /// This is equivalent of Node's `new https.Agent([options])`.
 HttpAgent createHttpsAgent([HttpAgentOptions options]) {
-  HTTPS https = require('https');
   var args = (options == null) ? [] : [options];
   return callConstructor(https.Agent, args);
 }
 
-/// Main entry point to Node's "http" module functionality.
-///
-/// Usage:
-///
-///     HTTP http = require('https');
-///     http.get("https://example.com");
-///
-/// See also:
-/// - [createHttpsAgent]
 @JS()
+@anonymous
 abstract class HTTPS {
   /// Returns a new instance of [HttpsServer].
   ///
@@ -59,4 +57,5 @@ abstract class HTTPS {
 }
 
 @JS()
+@anonymous
 abstract class HttpsServer extends TlsServer implements HttpServer {}