+++ lib/net.dart
@@ -1,12 +1,20 @@
 // Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
+
+/// NodeJS Net module.
+///
+/// Use top-level [net] object to access this module functionality.
 @JS()
-library node_interop.bindings.net;
+library node_interop.net;
 
 import 'package:js/js.dart';
 import 'events.dart';
+import 'node.dart';
+
+Net get net => require('net');
 
 @JS()
+@anonymous
 abstract class Net {
   external num isIP(String input);
   external bool isIPv4(String input);