+++ lib/stream.dart
@@ -1,10 +1,12 @@
-// Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
+// Copyright (c) 2018, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
+/// Node "stream" module bindings.
 @JS()
-library node_interop.bindings.stream;
+library node_interop.stream;
 
 import 'package:js/js.dart';
+
 import 'events.dart';
 
 @JS()