+++ lib/tls.dart
@@ -1,16 +1,17 @@
 // Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
+
+/// Node TLS module bindings.
 @JS()
-library node_interop.bindings.os;
+library node_interop.tls;
 
 import 'package:js/js.dart';
 
-@JS()
-abstract class OS {
-  external List<CPU> cpus();
-  external String tmpdir();
-  external String hostname();
-}
+import 'net.dart';
 
 @JS()
-abstract class CPU {}
+abstract class TlsServer extends NetServer {}
+
+@JS()
+@anonymous
+abstract class TlsServerOptions {}